Add pool events breakout files.
